# api/api/common_service.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "api/common_service.proto",
    "version": "version not set"
  },
  "schemes": [
    "http",
    "https"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/GetConfig": {
      "post": {
        "summary": "Dump Config in memory.",
        "operationId": "GetConfig",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/apiGetConfigResponse"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/apiGetConfigRequest"
            }
          }
        ],
        "tags": [
          "CommonService"
        ]
      }
    },
...
    "/Shutdown": {
      "post": {
        "summary": "Stop main process.",
        "operationId": "Shutdown",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/apiShutdownResponse"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/apiShutdownRequest"
            }
          }
        ],
        "tags": [
          "CommonService"
        ]
      }
    }
  },
  "definitions": {
...
    "apiGetConfigRequest": {
      "type": "object",
      "properties": {
        "header": {
          "$ref": "#/definitions/apiRequestHeader"
        }
      }
    },
    "apiGetConfigResponse": {
      "type": "object",
      "properties": {
        "header": {
          "$ref": "#/definitions/apiResponseHeader"
        },
        "message": {
          "type": "string"
        }
      }
    },
...
    "apiRequestHeader": {
      "type": "object",
      "properties": {
        "request_id": {
          "type": "string"
        },
        "start_time": {
          "type": "string",
          "format": "uint64"
        }
      }
    },
    "apiResponseHeader": {
      "type": "object",
      "properties": {
        "request_id": {
          "type": "string"
        },
        "resp_time": {
          "type": "string",
          "format": "uint64"
        },
        "ret_code": {
          "$ref": "#/definitions/apiRetCode"
        },
        "err_msg": {
          "type": "string"
        },
        "params": {
          "type": "object",
          "additionalProperties": {
            "type": "string"
          }
        }
      }
    },
    "apiRetCode": {
      "type": "string",
      "enum": [
        "OK",
        "FAIL",
        "ERROR"
      ],
      "default": "OK",
      "title": "- OK: Please do not modify bellow error codes!!!"
    },
    "apiShutdownRequest": {
      "type": "object",
      "properties": {
        "header": {
          "$ref": "#/definitions/apiRequestHeader"
        }
      }
    },
    "apiShutdownResponse": {
      "type": "object",
      "properties": {
        "header": {
          "$ref": "#/definitions/apiResponseHeader"
        },
        "message": {
          "type": "string"
        }
      }
    }
  }
}